Changed way of data-input to the version of chi2fitXYErr
The data has now to be put in by an external file, like in chi2fitXYErr.
This included also the way how the name of the output is specified.
There is still the old code included but as comment.
The way how to reactivate the old input is documented in the code.